# memory: optimize peer store memory footprint and performance