This applies the optimization I discovered in #50 to our `memory` implemetation.

Benchmark:
```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkPeerStore_PutSeeder-8                                       808           685           -15.22%
BenchmarkPeerStore_PutSeeder1KInfohash-8                             925           730           -21.08%
BenchmarkPeerStore_PutSeeder1KSeeders-8                              835           746           -10.66%
BenchmarkPeerStore_PutSeeder1KInfohash1KSeeders-8                    816           660           -19.12%
BenchmarkPeerStore_PutDeleteSeeder-8                                 2438          2160          -11.40%
BenchmarkPeerStore_PutDeleteSeeder1KInfohash-8                       2644          2613          -1.17%
BenchmarkPeerStore_PutDeleteSeeder1KSeeders-8                        2530          2489          -1.62%
BenchmarkPeerStore_PutDeleteSeeder1KInfohash1KSeeders-8              2604          2422          -6.99%
BenchmarkPeerStore_DeleteSeederNonExist-8                            472           403           -14.62%
BenchmarkPeerStore_DeleteSeederNonExist1KInfohash-8                  438           439           +0.23%
BenchmarkPeerStore_DeleteSeederNonExist1KSeeders-8                   458           420           -8.30%
BenchmarkPeerStore_DeleteSeederNonExist1KInfohash1KSeeders-8         462           417           -9.74%
BenchmarkPeerStore_PutGraduateDeleteLeecher-8                        3855          3340          -13.36%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KInfohash-8              3957          3779          -4.50%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KSeeders-8               4081          3726          -8.70%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KInfohash1KSeeders-8     4052          3679          -9.21%
BenchmarkPeerStore_GraduateLeecherNonExist-8                         916           802           -12.45%
BenchmarkPeerStore_GraduateLeecherNonExist1KInfohash-8               916           763           -16.70%
BenchmarkPeerStore_GraduateLeecherNonExist1KSeeders-8                889           774           -12.94%
BenchmarkPeerStore_GraduateLeecherNonExist1KInfohash1KSeeders-8      909           737           -18.92%
BenchmarkPeerStore_AnnouncePeers-8                                   1140          1091          -4.30%
BenchmarkPeerStore_AnnouncePeers1KInfohash-8                         1110          1152          +3.78%
BenchmarkPeerStore_AnnouncePeersSeeder-8                             909           874           -3.85%
BenchmarkPeerStore_AnnouncePeersSeeder1KInfohash-8                   999           854           -14.51%
BenchmarkPeerStore_GetSeeders-8                                      912           870           -4.61%
BenchmarkPeerStore_GetSeeders1KInfohash-8                            890           809           -9.10%
BenchmarkPeerStore_NumSeeders-8                                      480           247           -48.54%
BenchmarkPeerStore_NumSeeders1KInfohash-8                            497           246           -50.50%
```

Memory before:
```
Testing peer store "memory", Config: map[shards:1]...
1 infohashes...
        0 seeders,         0 leechers:             3816B (         3KiB)
        1 seeders,         0 leechers:             1448B (         1KiB)
        0 seeders,         1 leechers:             1448B (         1KiB)
        1 seeders,         1 leechers:             2472B (         2KiB)
        1 seeders,         1 leechers:             2472B (         2KiB)
        5 seeders,         5 leechers:             2792B (         2KiB)
       10 seeders,        10 leechers:             4728B (         4KiB)
       25 seeders,        25 leechers:            10152B (         9KiB)
       50 seeders,        50 leechers:            20984B (        20KiB)
      100 seeders,       100 leechers:            40216B (        39KiB)
      250 seeders,       250 leechers:           136184B (       132KiB)
      500 seeders,       500 leechers:           257176B (       251KiB)
     1000 seeders,      1000 leechers:           495576B (       483KiB)
     2500 seeders,      2500 leechers:          1058968B (      1034KiB,      1.0MiB)
     5000 seeders,      5000 leechers:          2111272B (      2061KiB,      2.0MiB)
    10000 seeders,     10000 leechers:          4252136B (      4152KiB,      4.1MiB)
    25000 seeders,     25000 leechers:          9629816B (      9404KiB,      9.2MiB)
    50000 seeders,     50000 leechers:         19284136B (     18832KiB,     18.4MiB)
   100000 seeders,    100000 leechers:         38462424B (     37560KiB,     36.7MiB)
   250000 seeders,    250000 leechers:        124646984B (    121725KiB,    118.9MiB)
   500000 seeders,    500000 leechers:        249360840B (    243516KiB,    237.8MiB)
  1000000 seeders,   1000000 leechers:        498840648B (    487149KiB,    475.7MiB)
  2500000 seeders,   2500000 leechers:       1073114488B (   1047963KiB,   1023.4MiB)
```
```
Testing peer store "memory", Config: map[shards:1]...
0 infohashes...
        1 seeders,         0 leechers:             3816B (         3KiB)

1 infohashes...
        1 seeders,         0 leechers:             1448B (         1KiB)

1 infohashes...
        1 seeders,         0 leechers:             1448B (         1KiB)

2 infohashes...
        1 seeders,         0 leechers:             2472B (         2KiB)

5 infohashes...
        1 seeders,         0 leechers:             5512B (         5KiB)

10 infohashes...
        1 seeders,         0 leechers:            10808B (        10KiB)

20 infohashes...
        1 seeders,         0 leechers:            21448B (        20KiB)

50 infohashes...
        1 seeders,         0 leechers:            52904B (        51KiB)

100 infohashes...
        1 seeders,         0 leechers:           105992B (       103KiB)

200 infohashes...
        1 seeders,         0 leechers:           211336B (       206KiB)

500 infohashes...
        1 seeders,         0 leechers:           537896B (       525KiB)

1000 infohashes...
        1 seeders,         0 leechers:          1073992B (      1048KiB,      1.0MiB)

2000 infohashes...
        1 seeders,         0 leechers:          2140184B (      2090KiB,      2.0MiB)

5000 infohashes...
        1 seeders,         0 leechers:          5309752B (      5185KiB,      5.1MiB)

10000 infohashes...
        1 seeders,         0 leechers:         10608776B (     10360KiB,     10.1MiB)

20000 infohashes...
        1 seeders,         0 leechers:         21219272B (     20721KiB,     20.2MiB)

50000 infohashes...
        1 seeders,         0 leechers:         52781016B (     51543KiB,     50.3MiB)

100000 infohashes...
        1 seeders,         0 leechers:        105568776B (    103094KiB,    100.7MiB)

200000 infohashes...
        1 seeders,         0 leechers:        211147624B (    206198KiB,    201.4MiB)

500000 infohashes...
        1 seeders,         0 leechers:        535711288B (    523155KiB,    510.9MiB)

1000000 infohashes...
        1 seeders,         0 leechers:       1071413000B (   1046301KiB,   1021.8MiB)
```

Memory after:
```
Testing peer store "memory", Config: map[shards:1]...
1 infohashes...
        0 seeders,         1 leechers:              872B
        1 seeders,         1 leechers:             1112B (         1KiB)
        5 seeders,         5 leechers:             1368B (         1KiB)
       10 seeders,        10 leechers:             2104B (         2KiB)
       25 seeders,        25 leechers:             4232B (         4KiB)
       50 seeders,        50 leechers:             7576B (         7KiB)
      100 seeders,       100 leechers:            15144B (        14KiB)
      250 seeders,       250 leechers:            44184B (        43KiB)
      500 seeders,       500 leechers:            90296B (        88KiB)
     1000 seeders,      1000 leechers:           182024B (       177KiB)
     2500 seeders,      2500 leechers:           385896B (       376KiB)
     5000 seeders,      5000 leechers:           772408B (       754KiB)
    10000 seeders,     10000 leechers:          1548648B (      1512KiB,      1.5MiB)
    25000 seeders,     25000 leechers:          3583064B (      3499KiB,      3.4MiB)
    50000 seeders,     50000 leechers:          7163464B (      6995KiB,      6.8MiB)
   100000 seeders,    100000 leechers:         14340552B (     14004KiB,     13.7MiB)
   250000 seeders,    250000 leechers:         43716344B (     42691KiB,     41.7MiB)
   500000 seeders,    500000 leechers:         87418200B (     85369KiB,     83.4MiB)
  1000000 seeders,   1000000 leechers:        174834392B (    170736KiB,    166.7MiB)
  2500000 seeders,   2500000 leechers:        389820904B (    380684KiB,    371.8MiB)
```
```
Testing peer store "memory", Config: map[shards:1]...
1 infohashes...
        1 seeders,         0 leechers:              872B

2 infohashes...
        1 seeders,         0 leechers:             1208B (         1KiB)

5 infohashes...
        1 seeders,         0 leechers:             2216B (         2KiB)

10 infohashes...
        1 seeders,         0 leechers:             4216B (         4KiB)

20 infohashes...
        1 seeders,         0 leechers:             8216B (         8KiB)

50 infohashes...
        1 seeders,         0 leechers:            20216B (        19KiB)

100 infohashes...
        1 seeders,         0 leechers:            39832B (        38KiB)

200 infohashes...
        1 seeders,         0 leechers:            79192B (        77KiB)

500 infohashes...
        1 seeders,         0 leechers:           209176B (       204KiB)

1000 infohashes...
        1 seeders,         0 leechers:           420696B (       410KiB)

2000 infohashes...
        1 seeders,         0 leechers:           831064B (       811KiB)

5000 infohashes...
        1 seeders,         0 leechers:          2008472B (      1961KiB,      1.9MiB)

10000 infohashes...
        1 seeders,         0 leechers:          4024280B (      3929KiB,      3.8MiB)

20000 infohashes...
        1 seeders,         0 leechers:          8043384B (      7854KiB,      7.7MiB)

50000 infohashes...
        1 seeders,         0 leechers:         19715000B (     19252KiB,     18.8MiB)

100000 infohashes...
        1 seeders,         0 leechers:         39450712B (     38526KiB,     37.6MiB)

200000 infohashes...
        1 seeders,         0 leechers:         78876920B (     77028KiB,     75.2MiB)

500000 infohashes...
        1 seeders,         0 leechers:        208508120B (    203621KiB,    198.8MiB)

1000000 infohashes...
        1 seeders,         0 leechers:        417069592B (    407294KiB,    397.7MiB)
```